% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_histogram_plot}
\alias{make_histogram_plot}
\title{Plot an histogram for a specified data frame column}
\usage{
make_histogram_plot(data = NULL, col = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{col}{A column of the `data` data frame that you wish to select.}
}
\value{
A plot
}
\description{
Plot an histogram for a specified data frame column
}
\details{
This function plots an histogram. It returns an error message if the given column name `col`is not correct.
See the examples below for its usage.
}
\examples{
# Return and show an histogram plot of the column "age" of the data frame "data"
plot <- make_histogram_plot(data, age)
plot
}
