% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_piechart_plot}
\alias{make_piechart_plot}
\title{Plot pie chart plot for a specified data frame column}
\usage{
make_piechart_plot(data = NULL, col = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{col}{A column of the `data` data frame that you wish to select.}
}
\value{
A plot
}
\description{
Plot pie chart plot for a specified data frame column
}
\details{
This function plots a pie chart. It returns an error message if the given column name `col`is not correct.
See the examples below for its usage.
}
\examples{
# Return and show a pie chart of the column "region" of the data frame "data".
plot <- make_piechart_plot(data, region)
plot
}
