% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_report}
\alias{make_report}
\title{create and show a plot for every column in the data frame (with some condition: see details)}
\usage{
make_report(data = NULL)
}
\arguments{
\item{data}{A data frame.}
}
\value{
Plots
}
\description{
create and show a plot for every column in the data frame (with some condition: see details)
}
\details{
if the column has less or exactly 5 number of distinct elements, it will show a pie chart
if the column has less or exactly between 6 and 50 number of distinct elements, it will show an histogram
otherwise it will show a density plot
it raises an error if the total number of column in the data frame is greater than 100, or if the given attributes `data` is not of a class "data.frame"
See the examples below for its usage.
}
\examples{
# create and show a plot for every column in the data frame `data`.
make_report(data)
}
