% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{unique_count}
\alias{unique_count}
\title{Count number of distinct elements in a specified data frame column}
\usage{
unique_count(data, col)
}
\arguments{
\item{data}{A data frame.}

\item{col}{A column of the `data` data frame that you wish to select.}
}
\value{
A `int`
}
\description{
Count number of distinct elements in a specified data frame column
}
\details{
This function allows compute the number of distinct elements in a data frame column. I doesn’t include NaN in the count.
See the examples below for its usage.
}
\examples{
# Return the number of unique elements in the column age for the data frame data

n <- unique_count(data, age)
}
