% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{unique_count}
\alias{unique_count}
\title{Count number of distinct elements in a specified column of a data frame}
\usage{
unique_count(data = NULL, col = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{col}{(type = "character"), a column of the data.frame `data` that you wish to select.}
}
\value{
An `integer`
}
\description{
Count number of distinct elements in a specified column of a data frame
}
\details{
This function computes the number of distinct elements in a data frame column.
See the examples below for its usage.
}
\examples{
# Return the number of unique elements in the column `age` of the data frame `data`.
n <- unique_count(data, "age")
}
